httpd_socket (inets v9.2)

Communication utility functions to be used by the Erlang web server API programmer.

This module provides the Erlang web server API module programmer with utility functions for generic sockets communication. The appropriate communication mechanism is transparently used, that is, ip_comm or ssl.

SEE ALSO

httpd

Functions

deliver(SocketType, Socket, IOListOrBinary)

deliver(SocketType, Socket, Data) -> Result

deliver/3 sends Data over Socket using the specified SocketType. Socket and SocketType is to be the socket and the socket_type form the mod record as defined in httpd.hrl

peername(SocketType, Socket)

peername(SocketType,Socket) -> {Port,IPAddress}

peername/2 returns the Port and IPAddress of the remote Socket.

resolve()

resolve() -> HostName

resolve/0 returns the official HostName of the current host.