httpd_socket (inets v7.4.1)

This module provides the Erlang web server API module programmer with utility functions for generic sockets communication. The appropriate communication mechanism is transparently used, that is, ip_comm or ssl.

SEE ALSO

httpd(3)

Functions

deliver(SocketType, Socket, Data) -> Result

SocketType = socket_type()

Socket = socket()

Data = io_list() | binary()

Result = socket_closed | void()

deliver/3 sends Data over Socket using the specified SocketType. Socket and SocketType is to be the socket and the socket_type form the mod record as defined in httpd.hrl

peername(SocketType,Socket) -> {Port,IPAddress}

SocketType = socket_type()

Socket = socket()

Port = integer()

IPAddress = string()

peername/2 returns the Port and IPAddress of the remote Socket.

resolve() -> HostName

HostName = string()

resolve/0 returns the official HostName of the current host.