inets (inets v9.2)

The Inets services API.

This module provides the most basic API to the clients and servers that are part of the Inets application, such as start and stop.

DATA TYPES

Type definitions that are used more than once in this module:

service() = httpc | httpd

property() = atom()

SEE ALSO

httpc, httpd

Functions

service_names()

service_names() -> [Service]

Returns a list of available service names.

services()

services() -> [{Service, Pid}]

Returns a list of currently running services.

Note

Services started as stand_alone are not listed.

services_info()

services_info() -> [{Service, Pid, Info}]

Returns a list of currently running services where each service is described by an [{Option, Value}] list. The information in the list is specific for each service and each service has probably its own info function that gives more details about the service. If specific service info returns {error, Reason}, Info will contain Reason term.

start()

Equivalent to start/1

start(Type)

start(Type) -> ok | {error, Reason}

Starts the Inets application. Default type is temporary. See also application.

start(Service, ServiceConfig)

Equivalent to start/3

start(Service, ServiceConfig, How)

start(Service, ServiceConfig, How) -> {ok, Pid} | {error, Reason}

Dynamically starts an Inets service after the Inets application has been started.

Note

Dynamically started services are not handled by application takeover and failover behavior when Inets is run as a distributed application. Nor are they automatically restarted when the Inets application is restarted. As long as the Inets application is operational, they are supervised and can be soft code upgraded.

A service started as stand_alone, that is, the service is not started as part of the Inets application, lose all OTP application benefits, such as soft upgrade. The stand_alone-service is linked to the process that started it. Usually some supervision functionality is still in place and in some sense the calling process becomes the top supervisor.

stop()

stop() -> ok

Stops the Inets application. See also application.

stop/2

stop(Service, Reference) -> ok | {error, Reason}

Stops a started service of the Inets application or takes down a stand_alone-service gracefully. When option stand_alone is used in start, only the pid is a valid argument to stop.