mod_auth (inets v7.4.1)

This module provides for basic user authentication using textual files, Dets databases, or Mnesia databases.

SEE ALSO

httpd(3), mod_alias(3)

Functions

add_group_member(GroupName, UserName, Options) -> true | {error, Reason}

GroupName = string()

UserName = string()

Options = [Option]

Option = {port,Port} | {addr,Address} | {dir,Directory} | {authPassword,AuthPassword}

Port = integer()

Address = {A,B,C,D} | string() | undefined

Dir = string()

AuthPassword = string()

Reason = term()

add_group_member/3, add_group_member/4, and add_group_member/5 each adds a user to a group. If the group does not exist, it is created and the user is added to the group. Upon successful operation, this function returns true. When add_group_members/3 is called, options Port and Dir are mandatory.

add_group_member(GroupName, UserName, Port, Dir) -> true | {error, Reason}

add_group_member(GroupName, UserName, Address, Port, Dir) -> true | {error, Reason}

add_user(UserName, Options) -> true| {error, Reason}

UserName = string()

Options = [Option]

Option = {password,Password} | {userData,UserData} | {port,Port} | {addr,Address} | {dir,Directory} | {authPassword,AuthPassword}

Password = string()

UserData = term()

Port = integer()

Address = {A,B,C,D} | string() | undefined

Dir = string()

AuthPassword =string()

Reason = term()

add_user/2, add_user/5, and add_user/6 each adds a user to the user database. If the operation is successful, this function returns true. If an error occurs, {error,Reason} is returned. When add_user/2 is called, options Password, UserData, Port, and Dir are mandatory.

add_user(UserName, Password, UserData, Port, Dir) -> true | {error, Reason}

add_user(UserName, Password, UserData, Address, Port, Dir) -> true | {error, Reason}

delete_group(GroupName, Options) -> true | {error,Reason} <name>delete_group(GroupName, Port, Dir) -> true | {error, Reason}

Options = [Option]

Option = {port,Port} | {addr,Address} | {dir,Directory} | {authPassword,AuthPassword}

Port = integer()

Address = {A,B,C,D} | string() | undefined

Dir = string()

GroupName = string()

AuthPassword = string()

Reason = term()

delete_group/2, delete_group/3, and delete_group/4 each deletes the group specified and returns true. If there is an error, {error, Reason} is returned. When delete_group/2 is called, option Port and Dir are mandatory.

delete_group(GroupName, Address, Port, Dir) -> true | {error, Reason}

delete_group_member(GroupName, UserName, Options) -> true | {error, Reason}

GroupName = string()

UserName = string()

Options = [Option]

Option = {port,Port} | {addr,Address} | {dir,Directory} | {authPassword,AuthPassword}

Port = integer()

Address = {A,B,C,D} | string() | undefined

Dir = string()

AuthPassword = string()

Reason = term()

delete_group_member/3, delete_group_member/4, and delete_group_member/5 each deletes a user from a group. If the group or the user does not exist, this function returns an error, otherwise true. When delete_group_member/3 is called, the options Port and Dir are mandatory.

delete_group_member(GroupName, UserName, Port, Dir) -> true | {error, Reason}

delete_group_member(GroupName, UserName, Address, Port, Dir) -> true | {error, Reason}

delete_user(UserName,Options) -> true | {error, Reason}

UserName = string()

Options = [Option]

Option = {port,Port} | {addr,Address} | {dir,Directory} | {authPassword,AuthPassword}

Port = integer()

Address = {A,B,C,D} | string() | undefined

Dir = string()

AuthPassword = string()

Reason = term()

delete_user/2, delete_user/3, and delete_user/4 each deletes a user from the user database. If the operation is successful, this function returns true. If an error occurs, {error,Reason} is returned. When delete_user/2 is called, options Port and Dir are mandatory.

delete_user(UserName, Port, Dir) -> true | {error, Reason}

delete_user(UserName, Address, Port, Dir) -> true | {error, Reason}

get_user(UserName,Options) -> {ok, #httpd_user} |{error, Reason}

UserName = string()

Options = [Option]

Option = {port,Port} | {addr,Address} | {dir,Directory} | {authPassword,AuthPassword}

Port = integer()

Address = {A,B,C,D} | string() | undefined

Dir = string()

AuthPassword = string()

Reason = term()

get_user/2, get_user/3, and get_user/4 each returns an httpd_user record containing the userdata for a specific user. If the user cannot be found, {error, Reason} is returned. When get_user/2 is called, options Port and Dir are mandatory.

get_user(UserName, Port, Dir) -> {ok, #httpd_user} | {error, Reason}

get_user(UserName, Address, Port, Dir) -> {ok, #httpd_user} | {error, Reason}

list_group_members(GroupName, Options) -> {ok, Users} | {error, Reason}

GroupName = string()

Options = [Option]

Option = {port,Port} | {addr,Address} | {dir,Directory} | {authPassword,AuthPassword}

Port = integer()

Address = {A,B,C,D} | string() | undefined

Dir = string()

Users = list()

AuthPassword = string()

Reason = term()

list_group_members/2, list_group_members/3, and list_group_members/4 each lists the members of a specified group. If the group does not exist or there is an error, {error, Reason} is returned. When list_group_members/2 is called, options Port and Dir are mandatory.

list_group_members(GroupName, Port, Dir) -> {ok, Users} | {error, Reason}

list_group_members(GroupName, Address, Port, Dir) -> {ok, Users} | {error, Reason}

list_groups(Options) -> {ok, Groups} | {error, Reason}

Options = [Option]

Option = {port,Port} | {addr,Address} | {dir,Directory} | {authPassword,AuthPassword}

Port = integer()

Address = {A,B,C,D} | string() | undefined

Dir = string()

Groups = list()

AuthPassword = string()

Reason = term()

list_groups/1, list_groups/2, and list_groups/3 each lists all the groups available. If there is an error, {error, Reason} is returned. When list_groups/1 is called, options Port and Dir are mandatory.

list_groups(Port, Dir) -> {ok, Groups} | {error, Reason}

list_groups(Address, Port, Dir) -> {ok, Groups} | {error, Reason}

list_users(Options) -> {ok, Users} | {error, Reason}
(since OTP R14B01)

Options = [Option]

Option = {port,Port} | {addr,Address} | {dir,Directory} | {authPassword,AuthPassword}

Port = integer()

Address = {A,B,C,D} | string() | undefined

Dir = string()

Users = list()

AuthPassword = string()

Reason = atom()

list_users/1, list_users/2, and list_users/3 each returns a list of users in the user database for a specific Port/Dir. When list_users/1 is called, options Port and Dir are mandatory.

list_users(Port, Dir) -> {ok, Users} | {error, Reason}
(since OTP R14B01)

list_users(Address, Port, Dir) -> {ok, Users} | {error, Reason}
(since OTP R14B01)

update_password(Port, Dir, OldPassword, NewPassword, NewPassword) -> ok | {error, Reason}

Port = integer()

Address = {A,B,C,D} | string() | undefined

Dir = string()

GroupName = string()

OldPassword = string()

NewPassword = string()

Reason = term()

update_password/5 and update_password/6 each updates AuthAccessPassword for the specified directory. If NewPassword is equal to "NoPassword", no password is required to change authorisation data. If NewPassword is equal to "DummyPassword", no changes can be done without changing the password first.

update_password(Address,Port, Dir, OldPassword, NewPassword, NewPassword) -> ok | {error, Reason}