mod_esi (inets v7.4.1)

This module defines the Erlang Server Interface (ESI) API. It is a more efficient way of writing Erlang scripts for your Inets web server than writing them as common CGI scripts.

DATA TYPES

The following data types are used in the functions for mod_esi:

env() =

{EnvKey()::atom(), Value::term()}

Currently supported key value pairs

{server_software, string()}

Indicates the inets version.

{server_name, string()}

The local hostname.

{gateway_interface, string()}

Legacy string used in CGI, just ignore.

{server_protocol, string()}

HTTP version, currently "HTTP/1.1"

{server_port, integer()}

Servers port number.

{request_method, "GET | "PUT" | "DELETE" | "POST" | "PATCH"}

HTTP request method.

{remote_adress, inet:ip_address()}

The clients ip address.

{peer_cert, undefined | no_peercert | DER:binary()}

For TLS connections where client certificates are used this will be an ASN.1 DER-encoded X509-certificate as an Erlang binary. If client certificates are not used the value will be no_peercert, and if TLS is not used (HTTP or connection is lost due to network failure) the value will be undefined.

{script_name, string()}

Request URI

{http_LowerCaseHTTPHeaderName, string()}

example: {http_content_type, "text/html"}

Functions

deliver(SessionID, Data) -> ok | {error, Reason}

SessionID = term()

Data = string() | io_list() | binary()

Reason = term()

This function is only intended to be used from functions called by the Erl Scheme interface to deliver parts of the content to the user.

Sends data from an Erl Scheme script back to the client.

If any HTTP header fields are added by the script, they must be in the first call to deliver/2, and the data in the call must be a string. Calls after the headers are complete can contain binary data to reduce copying overhead. Do not assume anything about the data type of SessionID. SessionID must be the value given as input to the ESI callback function that you implemented.