mod_security behaviour (inets v9.2)

Security Audit and Trailing Functionality

Callbacks

event(What, Port, Dir, Data)
(since OTP 18.1)

-callback event(What, Port, Dir, Data) -> term()
         when
             What :: auth_fail | user_block | user_unblock,
             Port :: integer(),
             Dir :: string(),
             Data :: [Info],
             Info :: {Name :: term(), Value :: term()}.

event(What, Address, Port, Dir, Data)
(since OTP 18.1)

-callback event(What, Address, Port, Dir, Data) -> term()
         when
             What :: auth_fail | user_block | user_unblock,
             Port :: integer(),
             Address :: inet:ip4_address() | string(),
             Dir :: string(),
             Data :: [Info],
             Info :: {Name :: term(), Value :: term()}.

event/4 or event/5 is called whenever an event occurs in the mod_security Erlang web server API module. (event/4 is called if Address is undefined, otherwise event/5. Argument What specifies the type of event that has occurred and is one of the following reasons:

auth_fail - A failed user authentication.

user_block - A user is being blocked from access.

user_unblock - A user is being removed from the block list.

Note

The event user_unblock is not triggered when a user is removed from the block list explicitly using the unblock_user function.

Functions

block_user(User, Port, Dir, Time)

Equivalent to block_user/5

block_user(User, Addr, Port, Dir, Time)

block_user(User, Address, Port, Dir, Seconds) -> true | {error, Reason}

block_user/4 and block_user/5 each blocks the user User from directory Dir for a specified amount of time.

list_auth_users(Port)

Equivalent to list_auth_users/3

list_auth_users/2

Equivalent to list_auth_users/3

list_auth_users(Addr, Port, Dir)

list_auth_users(Address, Port, Dir) -> Users | []

list_auth_users/1, list_auth_users/2, and list_auth_users/3 each returns a list of users that are currently authenticated. Authentications are stored for SecurityAuthTimeout seconds, and then discarded.

list_blocked_users(Port)

Equivalent to list_blocked_users/3

list_blocked_users/2

Equivalent to list_blocked_users/3

list_blocked_users(Addr, Port, Dir)

list_blocked_users(Address, Port, Dir) -> Users | []

list_blocked_users/1, list_blocked_users/2, and list_blocked_users/3 each returns a list of users that are currently blocked from access.

unblock_user(User, Port)

Equivalent to unblock_user/4

unblock_user/3

Equivalent to unblock_user/4

unblock_user(User, Addr, Port, Dir)

unblock_user(User, Address, Port, Dir) -> true | {error, Reason}

unblock_user/2, unblock_user/3, and unblock_user/4 each removes the user User from the list of blocked users for Port (and Dir).