mod_security (inets v7.4.1)

Security Audit and Trailing Functionality

Functions

block_user(User, Port, Dir, Seconds) -> true | {error, Reason}

User = string()

Port = integer()

Address = {A,B,C,D} | string() | undefined

Dir = string()

Seconds = integer() | infinity

Reason = no_such_directory

block_user/4 and block_user/5 each blocks the user User from directory Dir for a specified amount of time.

block_user(User, Address, Port, Dir, Seconds) -> true | {error, Reason}

list_auth_users(Port) -> Users | []

Port = integer()

Address = {A,B,C,D} | string() | undefined

Dir = string()

Users = list() = [string()]

list_auth_users/1, list_auth_users/2, and list_auth_users/3 each returns a list of users that are currently authenticated. Authentications are stored for SecurityAuthTimeout seconds, and then discarded.

list_auth_users(Port, Dir) -> Users | [] list_auth_users(Address, Port) -> Users | []

list_auth_users(Address, Port, Dir) -> Users | []

list_blocked_users(Port) -> Users | []

Port = integer()

Address = {A,B,C,D} | string() | undefined

Dir = string()

Users = list() = [string()]

list_blocked_users/1, list_blocked_users/2, and list_blocked_users/3 each returns a list of users that are currently blocked from access.

list_blocked_users(Port, Dir) -> Users | [] list_blocked_users(Address, Port) -> Users | []

list_blocked_users(Address, Port, Dir) -> Users | []

unblock_user(User, Port) -> true | {error, Reason}

User = string()

Port = integer()

Address = {A,B,C,D} | string() | undefined

Dir = string()

Reason = term()

unblock_user/2, unblock_user/3, and unblock_user/4 each removes the user User from the list of blocked users for Port (and Dir).

unblock_user(User, Port, Dir) -> true | {error, Reason} unblock_user(User, Address, Port) -> true | {error, Reason}

unblock_user(User, Address, Port, Dir) -> true | {error, Reason}